pyproject.toml
[project]
name = "chat-with-x-backend-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "chroma>=0.2.0",
    "dashscope>=1.24.0",
    "duckdb>=1.3.2",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.8",
    "langchain-openai>=0.3.28",
    "langchain-weaviate>=0.0.5",
    "langgraph>=0.5.4",
    "langsmith>=0.4.8",
    "lxml>=6.0.0",
    "protobuf==5.29.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "weaviate-cli>=3.2.3",
]

[dependency-groups]
dev = [
    "rich>=14.1.0",
    "jupyter>=1.1.1",
]
